feat(ME-1981): basic sftp support on windows
